backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
segmentation-models-pytorch==0.3.3
scikit-image>=0.21.0
tifffile>=2023.0.0
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
pydantic>=2.0.0
